This is an initial stab at porting the current Scarab code base to Turbine 2.4 and Yaafi (Yet Another Avalon Framework Implementation). To get it to build, you will probably have to copy the following libraries to your local Maven repository:
  www/repository/ecs/jars/ecs-1.4.2.jar
  www/repository/turbine/jars/turbine-2.4-dev.jar
  www/repository/turbine/jars/turbine-torque-impl-2.4-dev.jar
  www/repository/fulcrum/jars/fulcrum-crypto-1.0.5.jar
  www/repository/fulcrum/jars/fulcrum-intake-1.0.4-dev.jar
  www/repository/fulcrum/jars/fulcrum-template-1.0.2-dev.jar
  www/repository/fulcrum/jars/fulcrum-yaafi-1.0.4.jar

There are still major issues, but you can at least login. Much
work is left to do though.


git-svn-id: http://scarab.tigris.org/svn/scarab@9805 3c7165bb-d2ef-0310-b724-bafd9a86081f

+++ branches/TURBINE_2_4_PORT/src/java/org/tigris/scarab/om/AttributeGroup.java
@@ -137,7 +137,7 @@ public  class AttributeGroup
         Integer id = getModuleId();
         if ( id != null ) 
         {
-            module = ModuleManager.getInstance(id);
+            module = ModuleManager.getInstance(id.intValue());
         }
         
         return module;
