+++ branches/TURBINE_2_4_PORT/src/java/org/tigris/scarab/om/RModuleIssueType.java
@@ -125,7 +125,7 @@ public class RModuleIssueType
         Integer id = getModuleId();
         if ( id != null ) 
         {
-            module = ModuleManager.getInstance(id);
+            module = ModuleManager.getInstance(id.intValue());
         }
         
         return module;
